requirements.txt
click>=8.1.7
ffmpeg>=1.4
ffpb
librosa>=0.10.1
moviepy>=1.0.3
numba>=0.58.1
numpy>=1.26.3
opencv-python>=4.9.0.80
pexels-api-py>=0.0.5
pip>=23.3.2
pixabay>=0.0.5
requests>=2.31.0
scipy>=1.11.4
tinytag>=1.10.1
tqdm>=4.66.1
